refactor(annomate): changed file structure to better depict panes
