+++ src_new/tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest
-from core.state import DatasetState
+from core.dataset_state import DatasetState
 
 
 @pytest.fixture
