+++ src_new/tests/unit/test_io_controller.py
@@ -1,7 +1,7 @@
 import json
 import csv
 import pytest
-from core.state import DatasetState
+from core.dataset_state import DatasetState
 from models.dataset_model import DatasetTableModel
 from controllers.io_controller import IOController
